requirements.txt
pandas
fastapi
faiss-cpu
uvicorn
scikit-learn
google-api-python-client
dotenv
matplotlib